REFACTOR : option-package 관련 클래스 리팩터링 및 주석 수정하자
