+++ src/main/java/com/piaar_store_manager/server/domain/account_book/dto/AccountBookDefDto.java
@@ -1,4 +1,4 @@
-package com.piaar_store_manager.server.model.account_book.dto;
+package com.piaar_store_manager.server.domain.account_book.dto;
 
 import java.util.Date;
 import java.util.UUID;
