+++ src/main/java/com/piaar_store_manager/server/domain/account_book/dto/AccountBookJoinDto.java
@@ -1,13 +1,13 @@
-package com.piaar_store_manager.server.model.account_book.dto;
+package com.piaar_store_manager.server.domain.account_book.dto;
 
 import java.util.Date;
 import java.util.UUID;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.piaar_store_manager.server.domain.account_book.entity.AccountBookEntity;
 import com.piaar_store_manager.server.domain.user.dto.UserDefDto;
 import com.piaar_store_manager.server.domain.user.dto.UserGetDto;
 import com.piaar_store_manager.server.domain.user.entity.UserEntity;
-import com.piaar_store_manager.server.model.account_book.entity.AccountBookEntity;
 import com.piaar_store_manager.server.model.expenditure_type.dto.ExpenditureTypeDto;
 import com.piaar_store_manager.server.model.expenditure_type.entity.ExpenditureTypeEntity;
