+++ src/main/java/com/piaar_store_manager/server/domain/account_book/proj/AccountBookJoinProj.java
@@ -1,7 +1,7 @@
-package com.piaar_store_manager.server.model.account_book.proj;
+package com.piaar_store_manager.server.domain.account_book.proj;
 
+import com.piaar_store_manager.server.domain.account_book.entity.AccountBookEntity;
 import com.piaar_store_manager.server.domain.user.entity.UserEntity;
-import com.piaar_store_manager.server.model.account_book.entity.AccountBookEntity;
 import com.piaar_store_manager.server.model.expenditure_type.entity.ExpenditureTypeEntity;
 
 public interface AccountBookJoinProj {
